# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reddit Content Platform API",
...
    "version": "1.0.0",
    "contact": {
      "name": "Reddit Content Platform",
      "url": "https://github.com/your-username/reddit-content-platform",
      "email": "support@yourcompany.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "x-api-version": "1.0.0",
    "x-api-versions-supported": [
      "v1"
    ],
    "x-api-version-header": "X-API-Version",
    "x-api-media-type": "application/vnd.reddit-platform.v1+json"
  },
  "paths": {
...
    "/api/v1/posts/": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Get Posts",
        "description": "Get paginated list of posts with filtering and sorting options.\nOnly returns posts from keywords owned by the authenticated user.\n\n**Filtering Options:**\n- keyword_id: Filter by specific keyword\n- subreddit: Filter by subreddit name (partial match)\n- author: Filter by author name (partial match)\n- search: Search in post title and content\n- min_score/max_score: Filter by score range\n- date_from/date_to: Filter by post creation date range\n\n**Sorting Options:**\n- sort_by: Field to sort by (created_at, post_created_at, score, num_comments, title, author, subreddit)\n- sort_order: asc or desc",
        "operationId": "get_posts_api_v1_posts__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "keyword_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by keyword ID",
              "title": "Keyword Id"
            },
            "description": "Filter by keyword ID"
          },
          {
            "name": "subreddit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by subreddit",
              "title": "Subreddit"
            },
            "description": "Filter by subreddit"
          },
          {
            "name": "author",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by author",
              "title": "Author"
            },
            "description": "Filter by author"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search in title and content",
              "title": "Search"
            },
            "description": "Search in title and content"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort field (created_at, post_created_at, score, num_comments, title, author, subreddit)",
              "default": "created_at",
              "title": "Sort By"
            },
            "description": "Sort field (created_at, post_created_at, score, num_comments, title, author, subreddit)"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort order",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort order"
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum score filter",
              "title": "Min Score"
            },
            "description": "Minimum score filter"
          },
          {
            "name": "max_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum score filter",
              "title": "Max Score"
            },
            "description": "Maximum score filter"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter posts from this date (ISO format)",
              "title": "Date From"
            },
            "description": "Filter posts from this date (ISO format)"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter posts to this date (ISO format)",
              "title": "Date To"
            },
            "description": "Filter posts to this date (ISO format)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/{post_id}": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Get Post Detail",
        "description": "Get detailed information about a specific post, including comments.\nOnly returns post if it belongs to a keyword owned by the authenticated user.",
        "operationId": "get_post_detail_api_v1_posts__post_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/keyword/{keyword_id}": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Get Posts By Keyword",
        "description": "Get posts for a specific keyword.\nOnly returns posts if the keyword belongs to the authenticated user.",
        "operationId": "get_posts_by_keyword_api_v1_posts_keyword__keyword_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "keyword_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Keyword Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort field",
              "default": "created_at",
              "title": "Sort By"
            },
            "description": "Sort field"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort order",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/search/query": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Search Posts",
        "description": "Search posts by title and content.\nOnly searches in posts from keywords owned by the authenticated user.",
        "operationId": "search_posts_api_v1_posts_search_query_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Search query",
              "title": "Q"
            },
            "description": "Search query"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort field",
              "default": "created_at",
              "title": "Sort By"
            },
            "description": "Sort field"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort order",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/posts/stats/summary": {
      "get": {
        "tags": [
          "posts"
        ],
        "summary": "Get Post Statistics",
        "description": "Get statistics about user's posts.\nReturns summary information like total posts, average score, etc.",
        "operationId": "get_post_statistics_api_v1_posts_stats_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/blog/posts": {
      "get": {
        "tags": [
          "public-blog"
        ],
        "summary": "List Blog Posts",
        "description": "Get paginated list of published blog posts.",
        "operationId": "list_blog_posts_api_v1_blog_posts_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Posts per page",
              "default": 10,
              "title": "Size"
            },
            "description": "Posts per page"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category (keyword)",
              "title": "Category"
            },
            "description": "Filter by category (keyword)"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tag",
              "title": "Tag"
            },
            "description": "Filter by tag"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(latest|oldest|popular)$",
              "description": "Sort order",
              "default": "latest",
              "title": "Sort"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicBlogPostListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blog/posts/{slug}": {
      "get": {
        "tags": [
          "public-blog"
        ],
        "summary": "Get Blog Post By Slug",
        "description": "Get a specific blog post by slug.",
        "operationId": "get_blog_post_by_slug_api_v1_blog_posts__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicBlogPostDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PostDetailResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "keyword_id": {
            "type": "integer",
            "title": "Keyword Id"
          },
          "reddit_id": {
            "type": "string",
            "title": "Reddit Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "num_comments": {
            "type": "integer",
            "title": "Num Comments"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "subreddit": {
            "type": "string",
            "title": "Subreddit"
          },
          "post_created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Post Created At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/CommentResponse"
            },
            "type": "array",
            "title": "Comments",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "keyword_id",
          "reddit_id",
          "title",
          "author",
          "score",
          "num_comments",
          "url",
          "subreddit",
          "post_created_at",
          "created_at",
          "updated_at"
        ],
        "title": "PostDetailResponse",
        "description": "Detailed post response including comments."
      },
      "PostListResponse": {
        "properties": {
          "posts": {
            "items": {
              "$ref": "#/components/schemas/PostResponse"
            },
            "type": "array",
            "title": "Posts"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_prev": {
            "type": "boolean",
            "title": "Has Prev"
          }
        },
        "type": "object",
        "required": [
          "posts",
          "total",
          "page",
          "per_page",
          "total_pages",
          "has_next",
          "has_prev"
        ],
        "title": "PostListResponse",
        "description": "Paginated post list response."
      },
...
      "PublicBlogPostDetail": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Blog post title"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Markdown formatted content"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "URL-friendly slug"
          },
          "meta_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meta Description",
            "description": "SEO meta description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Content tags"
          },
          "featured_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Featured Image Url",
            "description": "Featured image URL"
          },
          "word_count": {
            "type": "integer",
            "title": "Word Count",
            "description": "Content word count"
          },
          "estimated_read_time": {
            "type": "integer",
            "title": "Estimated Read Time",
            "description": "Estimated read time in minutes"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "published_at": {
            "type": "string",
            "format": "date-time",
            "title": "Published At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "slug",
          "word_count",
          "estimated_read_time",
          "id",
          "published_at",
          "updated_at"
        ],
        "title": "PublicBlogPostDetail",
        "description": "Schema for detailed blog post view."
      },
      "PublicBlogPostListResponse": {
        "properties": {
          "posts": {
            "items": {
              "$ref": "#/components/schemas/PublicBlogPostSummary"
            },
            "type": "array",
            "title": "Posts"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_prev": {
            "type": "boolean",
            "title": "Has Prev"
          }
        },
        "type": "object",
        "required": [
          "posts",
          "total",
          "page",
          "size",
          "pages",
          "has_next",
          "has_prev"
        ],
        "title": "PublicBlogPostListResponse",
        "description": "Schema for paginated blog post list."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/login endpoint"
      },
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://www.reddit.com/api/v1/authorize",
            "tokenUrl": "/api/v1/auth/token",
            "scopes": {
              "read": "Read Reddit posts and comments",
              "identity": "Access user identity information"
            }
          }
        },
        "description": "Reddit OAuth2 authentication"
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://your-app.railway.app",
      "description": "Production server"
    }
  ],
...
}